refactor: rename referenceImplementation.ts to shuffleReference.ts
